% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_permutation.R
\name{build_permutation}
\alias{build_permutation}
\title{build_permuation}
\usage{
build_permutation(df1, df2, p)
}
\arguments{
\item{df1}{The first dataframe.}

\item{df2}{The second dataframe.}

\item{p}{The permutation to apply.}
}
\description{
Apply a permuation to obtain a full linked data set from df1 and df2. 
The permutation is applied to df2, reordering its rows with respect to df1.
}
\details{
This file is part of GFS.

GFS is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

GFS is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with GFS. If not, see <https://www.gnu.org/licenses/>.
}
