% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df1.R
\docType{data}
\name{df1}
\alias{df1}
\title{Partition 1 of simulated example data set for reference.}
\format{
A data frame with columns:
\describe{
  See df_full for information about how the data set was generated.
}
}
\usage{
df1
}
\description{
Partition 1 of simulated example data set for reference.
}
\examples{
\dontrun{
 df1
}
}
\keyword{datasets}
