% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_inputs.R
\name{permute_inputs}
\alias{permute_inputs}
\title{Runs the permutation sampling procedure on the two datasets specified by the
df1_path and df2_path inputs over the blocks defined in the datasets.}
\usage{
permute_inputs(df1_path, df2_path, formula, family, N, I, t, burn_in,
  sample_interval, conda_env = "NA", activate_env = "NA",
  keep_csv = FALSE)
}
\arguments{
\item{df1_path}{A string with the path (from the R working directory) to the
first dataset with blocks.}

\item{df2_path}{A string with the path (from the R working directory) to the
second dataset with blocks.}

\item{formula}{A string with the formula specifying the response variable
and the covariates. Written in the R format (e.g. 'y ~ x1 + x2 + x3').}

\item{family}{A string specifying the family of distributions used in the
    regression step. Current supported families are 'Normal', 'Logistic',
"Poisson".}

\item{N}{An integer specifying the number of desired full permutations. The
number of iterations will be burn_in + N*sample_interval.}

\item{I}{An integer specifying how many iterations to complete in sampling
the regression parameters.}

\item{t}{An integer specifying the number of iterations (successful or not)
in the Metropolis-Hastings step.}

\item{burn_in}{An integer specifying the number of full iterations to be
completed before any samples are taken.}

\item{sample_interval}{An integer specifying the number of full iterations
in between each sample.}

\item{conda_env}{'NA' by default. If using the Anaconda distribution of
python, specify which environment to use. Runs 'source activate conda_env'.}

\item{activate_env}{'NA' by default. Enter a string with the command used to
activate the desired python environment if it neither the default
environment nor an Anaconda environment.}

\item{keep_csv}{FALSE by default. If TRUE, the .csv file containing the
permutations will not be automatically deleted upon completion of the
procedure.}
}
\value{
A dataframe with number of rows equal to the number of rows in the
			first dataset and N columns. Each column is a full permutation with
			respect to the first dataset to be applied to the second dataset.
}
\description{
Runs the permutation sampling procedure on the two datasets specified by the
df1_path and df2_path inputs over the blocks defined in the datasets.
}
