% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/py_setup.R
\name{py_setup}
\alias{py_setup}
\title{Installs the required python files to the package directory.
No input is required, but the force argument can be set. 
Python code can be found on GitHub at:
https://github.com/edwinfarley/GFSpython}
\usage{
py_setup(force = FALSE)
}
\arguments{
\item{force}{Whether to remove existing python directory and re-install components.}
}
\description{
Installs the required python files to the package directory.
No input is required, but the force argument can be set. 
Python code can be found on GitHub at:
https://github.com/edwinfarley/GFSpython
}
