% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples1.R
\docType{data}
\name{samples1}
\alias{samples1}
\title{Partition 1 of example subset of the Behavioral risk factor surveillance system survey data set (BRFSS)
consisting of 3200 rows extracted from the full data set. Entries sampled at random from
complete rows (no missing or warning values).}
\format{
A data frame with columns:
\describe{
 \item{PHYSHLTH}{A measure of physical health on a scale of 1 to 30.}
 \item{MENTHLTH}{A measure of mental health on a scale of 1 to 30.}
 \item{AGE}{The respondent's age.}
 \item{WEIGHT}{The respondent's weight in lbs.}
 \item{block}{Blocking index.}
}
}
\source{
Department of Health, Centers for Disease Control Human Services, and Pre-
vention (CDC). Behavioral risk factor surveillance system survey data, 2001.
}
\usage{
samples1
}
\description{
Partition 1 of example subset of the Behavioral risk factor surveillance system survey data set (BRFSS)
consisting of 3200 rows extracted from the full data set. Entries sampled at random from
complete rows (no missing or warning values).
}
\examples{
\dontrun{
 samples1
}
}
\keyword{datasets}
