% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples2.R
\docType{data}
\name{samples2}
\alias{samples2}
\title{Partition 2 of example subset of the Behavioral risk factor surveillance system survey data set (BRFSS)
consisting of 3200 rows extracted from the full data set. Entries sampled at random from
complete rows (no missing or warning values).}
\format{
A data frame with columns:
\describe{
 \item{GENHLTH}{A measure of general health on a scale of 1 to 5.}
 \item{ALCDAY}{The number of days a year the respondent drinks alcohol.}
 \item{ASTHMA}{Whether the respondent has asthma.}
 \item{block}{Blocking index.}
}
}
\source{
Department of Health, Centers for Disease Control Human Services, and Pre-
vention (CDC). Behavioral risk factor surveillance system survey data, 2001.
}
\usage{
samples2
}
\description{
Partition 2 of example subset of the Behavioral risk factor surveillance system survey data set (BRFSS)
consisting of 3200 rows extracted from the full data set. Entries sampled at random from
complete rows (no missing or warning values).
}
\examples{
\dontrun{
 samples2
}
}
\keyword{datasets}
