% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_permutation.R
\name{switch_permutation}
\alias{switch_permutation}
\title{Switch the direction of the permutation to be in terms of the data set. The
permute_inputs function returns a permutation in terms of the second data set
with respect to the first data set. This function will change the permutation
to be in terms of the first data set with respect to the second.}
\usage{
switch_permutation(P)
}
\arguments{
\item{P}{Dataframe/matrix of permutations as columns.}
}
\description{
Switch the direction of the permutation to be in terms of the data set. The
permute_inputs function returns a permutation in terms of the second data set
with respect to the first data set. This function will change the permutation
to be in terms of the first data set with respect to the second.
}
