% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_P.R
\docType{data}
\name{true_P}
\alias{true_P}
\title{true_P}
\format{
A data frame with columns:
\describe{
  Each column is a permutation.
}
}
\usage{
true_P
}
\description{
The true permuation linking samples1 and samples2.
}
\details{
This file is part of GFS.

GFS is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

GFS is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with GFS. If not, see <https://www.gnu.org/licenses/>.
}
\examples{
\dontrun{
 true_P
}
}
\keyword{datasets}
