% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_permutation.R
\name{build_permutation}
\alias{build_permutation}
\title{Apply a permuation to obtain a full linked data set from df1 and df2. 
The permutation is applied to df2, reordering its rows with respect to df1.}
\usage{
build_permutation(df1, df2, p)
}
\arguments{
\item{df1}{The first dataframe.}

\item{df2}{The second dataframe.}

\item{p}{The permutation to apply.}
}
\description{
Apply a permuation to obtain a full linked data set from df1 and df2. 
The permutation is applied to df2, reordering its rows with respect to df1.
}
