% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2.R
\docType{data}
\name{df2}
\alias{df2}
\title{Partition 2 of simulated example data set for reference.}
\format{
A data frame with columns:
\describe{
  See df_full for information about how the data set was generated.
}
}
\usage{
df2
}
\description{
Partition 2 of simulated example data set for reference.
}
\examples{
\dontrun{
 df2
}
}
\keyword{datasets}
