% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_inputs.R
\name{permute_inputs}
\alias{permute_inputs}
\title{Runs the permutation sampling procedure on the two data sets specified by the
df1_path and df2_path inputs over the blocks defined in the data sets. A
sampled permutation can be applied to the data set at df2_path to aling its
rows with those of the data set at df1_path.}
\usage{
permute_inputs(
  df1_path,
  df2_path,
  formula,
  family,
  N,
  I,
  t,
  burn_in,
  sample_interval,
  block_name = "block",
  conda_env = "NA",
  activate_env = "NA",
  python = "python"
)
}
\arguments{
\item{df1_path}{A string with the path (from the R working directory) to the
first data set with blocks.}

\item{df2_path}{A string with the path (from the R working directory) to the
   second data set with blocks. This data set MUST contain the response
variable from the formula.}

\item{formula}{A string with the formula specifying the response variable
and the covariates. Written in the R format (e.g. 'y ~ x1 + x2 + x3').}

\item{family}{A string specifying the family of distributions used in the
    regression step. Current supported families are 'Normal', 'Logistic',
"Poisson".}

\item{N}{An integer specifying the number of desired full permutations. The
number of iterations will be burn_in + N*sample_interval.}

\item{I}{An integer specifying how many iterations to complete in sampling
the regression parameters.}

\item{t}{An integer specifying the number of iterations (successful or not)
in the Metropolis-Hastings step.}

\item{burn_in}{An integer specifying the number of full iterations to be
completed before any samples are taken.}

\item{sample_interval}{An integer specifying the period of sampling. If 
sample_interval is 1, every full iteration will be a sample, if
sample_interval is 2, every other full iteration will be a sample,
 and so on.}

\item{block_name}{The name of the column in the two input data sets that
contains the number of the block to which each row belongs. Set to
"block" by default. Column name must be the same in the two data sets.}

\item{conda_env}{'NA' by default. If using the Anaconda distribution of
python, specify which environment to use. Runs 'source activate conda_env'.}

\item{activate_env}{'NA' by default. Enter a string with the command used to
activate the desired python environment if it is neither the default
environment nor an Anaconda environment.}

\item{python}{The name of the installed python deployment. If a conda environment
was set up with create_conda_environment, this should be "python".}
}
\value{
A dataframe with number of rows equal to the number of rows in the
			first data set and N columns. Each column is a full permutation with
			respect to the first data set to be applied to the second data set.
}
\description{
Runs the permutation sampling procedure on the two data sets specified by the
df1_path and df2_path inputs over the blocks defined in the data sets. A
sampled permutation can be applied to the data set at df2_path to aling its
rows with those of the data set at df1_path.
}
