% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples2.R
\docType{data}
\name{samples2}
\alias{samples2}
\title{samples2}
\format{
A data frame with columns:
\describe{
 \item{GENHLTH}{A measure of general health on a scale of 1 to 5.}
 \item{ALCDAY}{The number of days a year the respondent drinks alcohol.}
 \item{ASTHMA}{Whether the respondent has asthma.}
 \item{block}{Blocking index.}
}
}
\usage{
samples2
}
\description{
Partition 2 of example subset of the Behavioral Risk Factor Surveillance System Survey (BRFSS) data set
consisting of 3200 rows extracted from the full data set. Entries sampled at random from
complete rows (no missing or warning values).
}
\details{
This file is part of GFS.

GFS is free software: you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation, either version 3 of the License, or
(at your option) any later version.

GFS is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU General Public License for more details.

You should have received a copy of the GNU General Public License
along with GFS. If not, see <https://www.gnu.org/licenses/>.
}
\examples{
\dontrun{
 samples2
}
}
\references{
Department of Health, Centers for Disease Control Human Services, and Pre-
vention (CDC). Behavioral risk factor surveillance system survey data, 2001.
}
\keyword{datasets}
